requirements.txt
# Backend Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# RAG & Embeddings
chromadb==0.4.18
sentence-transformers==2.2.2

# LLM Interface
llama-cpp-python==0.2.20

# Utilities
jinja2==3.1.2
markdown==3.5.1
python-dotenv==1.0.0

# Optional - Si besoin de télécharger modèles
huggingface-hub==0.19.4

# Déjà installé
PyMuPDF==1.23.8
